enb_stack_lte: implement the PDCP interface for GTPU on the stack
this prepares for a change in which GTPU no longer writes into PDCP
directly but instead uses the stack as a wrapper.

For this the interface will be changed to use the EPS bearer ID instead
of the LCID

The stack will know which PDCP entity (EUTRA or NR) is currently associated
with the EPS bearer ID and will forward the PDU accordingly.