sched,nr: add test case for SSB scheduling
The SSB test checks whether, with the given periodicity, the
SSB grant is present within the dl_sched structure passed to the PHY.
It also checks that the SSB grant is not scheduled outside
the required time slots.

NOTE: The test is a modified version of an existing scheduler test.

Signed-off-by: Carlo Galiotto <carlo@srs.io>