emergency_handlers: increase max handlers to 256
in the ttcn3_dut application many "virtuaL" UE instances are created and
destroyed when executing the tests.

With the previous limit of 12 handlers the application stopped after
a few tests. With the limit raised to 256 we should be able to run
all white_listed TTCN3 tests for the UE without problems.

This fixes #3886