ue,mac_nr: reset DL and UL HARQ when MAC is reset
DL and UL harq entities where never reset when the UE was going into IDLE.
As a result, the first transmissions where always using the old softbuffer
without resetting it, resulting in wrong TB being decoded.

In the case of the RAR, the UE decoded a wrong (in fact the old) TCRNTI
and the RA procedure failed.